# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Social Network API",
    "description": "HTTP contract for the Go backend and React JavaScript frontend. Posting supports privacy, nested comments, JPEG/PNG/GIF media, mutually exclusive like/dislike votes, author edits, soft-deletion tombstones, groups, events, chat, notifications, and WebSocket delivery.",
    "version": "1.3.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session_token"
      }
    },
...
    "parameters": {
      "ResourceId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Unique identifier of the resource addressed by this endpoint."
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        },
        "description": "Maximum number of top-level resources to return."
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request",
        "content": {
          "application/json": {
            "schema": {
              "allOf": [
                {
                  "$ref": "#/components/schemas/Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "const": "error",
                      "description": "Indicates that the request failed."
                    }
                  }
                }
              ]
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/users/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ResourceId"
        }
      ],
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get a user's profile",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "description": "Returns the selected user's complete profile when it is public, owned by the requester, or visible through an accepted follow relationship."
      }
    },
...
    "/groups/{id}/invite": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "post": {
        "tags": [
          "Groups"
        ],
        "summary": "Invite a user to a group",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "user_id": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invitation sent"
          }
        },
        "description": "Invites an eligible user to join the selected group."
      }
    },
...
    "/users/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "first_name",
                  "last_name",
                  "date_of_birth"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "date_of_birth": {
                    "type": "string",
                    "format": "date"
                  },
                  "avatar": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "nickname": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "about_me": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "is_public": {
                    "type": "boolean",
                    "default": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "description": "Creates a user account from the submitted profile and credentials."
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create a session",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session cookie set"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "description": "Authenticates credentials and creates the session cookie used by protected endpoints."
      }
    },
...
    "/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get the current user",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "description": "Returns the authenticated user's complete profile and account settings."
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update the current user's profile",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "current_password": {
                    "type": "string",
                    "description": "Required when changing the email address or password."
                  },
                  "new_password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "date_of_birth": {
                    "type": "string",
                    "format": "date"
                  },
                  "nickname": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "description": "The optional username or display name."
                  },
                  "about_me": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "is_public": {
                    "type": "boolean"
                  },
                  "avatar": {
                    "type": "string",
                    "format": "binary",
                    "description": "Replacement JPEG, PNG, or GIF avatar."
                  },
                  "remove_avatar": {
                    "type": "boolean",
                    "default": false,
                    "description": "Removes the current avatar when true. Cannot be combined with avatar."
                  }
                },
                "minProperties": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated current user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrentUser"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "description": "Updates any supplied account or profile fields. At least one field is required; changing email or password requires the current password."
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete the current user's account",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "password"
                ],
                "properties": {
                  "password": {
                    "type": "string",
                    "description": "The current password used to confirm permanent account deletion."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account deleted and session cookie cleared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Response"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        },
        "description": "Permanently deletes the account in one transaction. Authored posts and comments are soft-deleted first and remain as tombstones; memberships, follows, sessions, votes, invitations, direct-message threads, messages, and notifications are removed. Groups and events created by the user remain without a creator."
      }
    },
    "/users/{id}/posts": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ResourceId"
        },
        {
          "$ref": "#/components/parameters/Limit"
        },
        {
          "$ref": "#/components/parameters/Offset"
        }
      ],
      "get": {
        "tags": [
          "Users",
          "Posts"
        ],
        "summary": "Get posts visible on a user's profile",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Visible profile posts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "description": "Returns posts the current user may view on the selected user's profile."
      }
    },
    "/posts": {
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Get the home feed or a group feed",
        "description": "Without group_id, excludes group posts and applies ordinary privacy. With group_id, requires accepted membership and returns only that group's posts.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          },
          {
            "name": "group_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "When supplied, returns the selected group's feed and requires accepted membership."
          }
        ],
        "responses": {
          "200": {
            "description": "Visible posts ordered newest first",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Posts"
        ],
        "summary": "Create a post",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "privacy": {
                    "type": "string",
                    "enum": [
                      "public",
                      "almost_private",
                      "private"
                    ],
                    "default": "public"
                  },
                  "group_id": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "uuid"
                  },
                  "audience_ids": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "description": "Repeat the audience_ids form field for each selected accepted follower."
                  },
                  "image": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "binary",
                    "description": "Optional JPEG, PNG, or GIF up to 5 MiB."
                  }
                },
                "anyOf": [
                  {
                    "required": [
                      "content"
                    ]
                  },
                  {
                    "required": [
                      "image"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "description": "Creates a profile, follower-only, selected-follower, or group post with optional media."
      }
    },
    "/posts/{id}": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ResourceId"
        }
      ],
      "get": {
        "tags": [
          "Posts"
        ],
        "summary": "Get a visible post or its tombstone",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Post detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        },
        "description": "Returns the post when visible, using a minimal tombstone if it has been soft-deleted."
      },
      "patch": {
        "tags": [
          "Posts"
        ],
        "summary": "Edit an active post as its author",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string"
                  },
                  "privacy": {
                    "type": "string",
                    "enum": [
                      "public",
                      "almost_private",
                      "private"
                    ]
                  },
                  "audience_ids": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uuid"
                    }
                  },
                  "image": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "format": "binary"
                  },
                  "remove_image": {
                    "type": "boolean",
                    "default": false
                  }
                },
                "minProperties": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated post",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          }
        },
        "description": "Updates the mutable content, media, and audience fields of the author's active post."
      },
      "delete": {
        "tags": [
          "Posts"
        ],
        "summary": "Soft-delete a post as its author",
        "description": "Idempotent. Comments remain readable to users authorized by the retained post visibility data, but the deleted post accepts no new comments or votes.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Minimal post tombstone",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/posts/{id}/vote": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ResourceId"
        }
      ],
      "put": {
        "tags": [
          "Votes"
        ],
        "summary": "Set or switch the current user's post vote",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "vote"
                ],
                "properties": {
                  "vote": {
                    "type": "string",
                    "enum": [
                      "like",
                      "dislike"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Current counts and vote",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        },
        "description": "Creates or replaces the current user's mutually exclusive like or dislike vote."
      },
      "delete": {
        "tags": [
          "Votes"
        ],
        "summary": "Remove the current user's post vote",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current counts with viewer_vote = none",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        },
        "description": "Removes the current user's vote and returns the updated totals."
      }
    },
...
    "/ws": {
      "get": {
        "tags": [
          "Realtime"
        ],
        "summary": "Open the realtime WebSocket connection",
        "description": "Upgrades an authenticated HTTP request to a WebSocket connection. The session_token cookie authenticates the user. Client-to-server payloads use JSON, for example `{ \"type\": \"message.send\", \"chat_id\": \"f9b81580-91f1-42ef-b0c9-f9170587d57c\", \"content\": \"Hello\" }`. Server-to-client message events look like `{ \"type\": \"message.created\", \"data\": <Message> }`; notification events look like `{ \"type\": \"notification.created\", \"data\": <Notification> }`; read receipts may use `{ \"type\": \"notification.read\", \"data\": { \"id\": \"b4dfcf9c-d761-48a2-9851-8b79974c9e20\" } }`. REST endpoints remain the source of truth for pagination and recovery after reconnects.",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "101": {
            "description": "WebSocket protocol upgrade accepted"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  }
}